Renamed Buttons to ApiExamples
