+++ web_frontend/src/App.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import LoginPage from './LoginPage';
 import DashboardPage from './DashboardPage';
-import Example from './ApiExamples';
-import Buttons from './Buttons';
+import ApiExamples from './ApiExamples';
 
 function App() {
     return (
@@ -12,8 +11,7 @@ function App() {
                 <Routes>
                     <Route path="/" element={<LoginPage />} />
                     <Route path="/DashboardPage" element={<DashboardPage />} />
-                    <Route path="/ApiExamples" element={<Example/>} />
-                    <Route path="/Buttons" element={<Buttons/>} />
+                    <Route path="/ApiExamples" element={<ApiExamples/>} />
                 </Routes>
             </div>
         </Router>
